Fix behavior of APISession.request w/r/t retry
This update fixes a few regression bugs in our internal tooling that only came
up as of recently when we contended with some non-transient 503s in the
Schedules API ;)

The documented and appropros behavior with respect to the retry logic is to
return the response object if retrying fails. This is because raising
exceptions is the task of higher-level functions i.e. the r{VERB} methods and
iter_all.

Rather, the lower-level client methods get, post, put and delete are expected
to return response objects, and any implementation that depended on this
behavior previously would have had to implement new exception handling
(try/catch) to avoid breakage.