Change how routing key is included in event payloads
Including the routing key via the `send_event` method broke some downstream usage of the client, which depended on being able to call `post` directly with an explicitly-defined payload that lacked a routing key (which was possible when using the header).

To remedy this, the `post` method is overridden in EventsAPIClient to add the `routing_key` parameter to the body before sending.