Increment offset by length of results in iter_all
The response plays API has been returning null values for offset and limit in its body. By trusting these values from the API, the client set itself up for a TypeError (see #61).

This fixes the issue by incrementing offset by the length of the array of results returned by the index instead of the value reported by the API.