Change error class in APISession.iter_cursor
PDClientError instead of the more generic ValueError is appropriate because it
follows the design of entity wrapping resources more closely.

This will potentially be a breaking change although the risk is low at this
stage. The only endpoints where this method could be used (that support
cursor-based pagination) do not require any fancy guesses. Only if another
cursor-paged index were added that departed from audit records APIs' pattern of
entity wrapping would there need to be any error handling for the error in
question, and such error handling is not recommended in the first place
(because it interferes with a very fundamental exception class that could be
indicative of a completely different problem elsewhere.