Simplify the test process
There are many different steps and aliases because (I think I remember it being this way) to support Python 2.7 we needed to do things a bit differently.

This is no longer needed and so the steps can be combined to make tests faster and make it fewer lines to add support for a new Python version.